+++ database/migrations/2020_04_18_170029_create_order_details_table.php
@@ -21,9 +21,9 @@ class CreateOrderDetailsTable extends Migration
             $table->text('write');
             $table->integer('note_status');
             $table->text('note');
-            $table->float('price');            
+            $table->decimal('price',8,2);            
             $table->integer('quantity')->default('1');
-            $table->float('sum_price');
+            $table->decimal('sum_price',8,2);
             $table->integer('upload_image_status');
             $table->integer('permission_to_publish')->nullable()->default(1);
             $table->integer('status')->default('1');
